Fix production vscode builds accessing stylex css bundle
Summary:
In the previous diff we added support for stylex in ISL. This works everywhere except for production builds of the vscode extension.

In that case, it does actually compile the stylex stuff into styles and even puts the classnames correctly. It just can't load the actual `assets/stylex.a1b2c3d4.css` file. That's because it is injected into the HTML, but our production webview defines its own HTML.

Instead, we need to add our own manual import of the stylex css file. But the problem here is that the vite plugin only outputs to a hardcoded place and always includes the hash.

So we need a tiny plugin that rewrites this file to match our others: `webview/res/stylex.css`, with no hash. Then we can add this file to our html defined in the extension.

This solution is a little complicated since we need a whole custom plugin. But it removes the need for us to do anything fancier with the html.

Reviewed By: quark-zju

Differential Revision: D53324215

fbshipit-source-id: 895f793ce438e7671229ca6a739dd8b0ea3aea5d