Component Explorer for debugging styles
Summary:
Add a simple component explorer we can use as a playground for colors and components and stylex and as a reference when designing new components. Also, if we migrate away from vscode webview ui toolkit, this will be an easy place to test appearance differences.

It's hidden behind the debug tools menu, so most users won't stumble across this.

It's dynamically imported, so this won't be part of the main bundle. We could go further to not even bundle this in production but I don't do that here.

Reviewed By: quark-zju

Differential Revision: D53237326

fbshipit-source-id: 6d0cfd694f1733d15085b21ef8da8a0616044f98