Add atomWithRefresh util
Summary:
`atomWithRefresh` is a pattern described by the jotai docs:
https://jotai.org/docs/recipes/atom-with-refresh

This is a nice pattern for working with a simple async function to fetch data, which is common in our app. For example, shelved changes. Each time you open the dropdown, it does a fetch.

I added a variant `atomLoadableWithRefresh` too, which is like `atomWithRefresh` but doesn't suspend. I found this a little easier to work with. Also the jotai `loadable` is nicer than the recoil one (checking the `.type` allows typescript to narrow so you don't need `valueOrThrow()` or `valueMaybe()`.

Note that we can't wrap `atomWithRefresh` with `loadable()` (as you typically can do: https://jotai.org/docs/utilities/async#loadable), since we want to access the refresh() function without checking if the data is loading.

Reviewed By: quark-zju

Differential Revision: D53324780

fbshipit-source-id: f18153177e801b6e18209e14e00a7541e2827c0b