commands: get rid of mutable config
Summary:
The only remaining user of the mutable ConfigSet was the "no repo" commands (that accepted a `&mut ConfigSet`). Change these instead to take `&Arc<dyn configmodel::Config)>`. I went with a `&Arc` instead of `Arc` since the former can be passed directly as a `&dyn Config`.

I added a ConfigSet::wrap() constructor that creates a mutable config by wrapping a non-mutable config. This was needed for a few spots that actually were inserting into the config for various reasons.

This change also forced me to change a lot of `&ConfigSet` parameters into `&dyn Config` (which is good).

I'm getting rid of the mutable config because:
  1. We shouldn't be mutating the config. If you really need to override config values at rutime, you should wrap the config.
  2. I want to put the config inside the CoreContext for convenience. To do that, I need an `Arc<dyn Config>` (and I didn't want to clone/diverge the ConfigSet).

Reviewed By: quark-zju

Differential Revision: D53250253

fbshipit-source-id: 1bb4072341a2ec70fac20607de2c68034380553d