clidispatch: set cli flag derived configs earlier
Summary:
There are some config values we set based on CLI flags. For example, if you pass "--quiet", we set "ui.quiet=true" in the config.

Previously we would load the repo, then mutate the repo's config to add in the derived values. I'm trying to get rid of this config mutability, so now instead we pass these derved configs earlier as inputs to config/repo loading (along with "--config" and "--configfile") values.

Reviewed By: quark-zju

Differential Revision: D53250255

fbshipit-source-id: 948887609f928255da1487997c9d89635a2523ff