configloader: generalize config overrides passed to hg::load()
Summary:
Previously we only supported passing the "--config" and "--configfile" CLI values. Now we generally support any raw (i.e. unparsed) configs, key/value configs, and config files. We also allow the caller to specify the config source.

I'm going to use this to pass the config overrides set during command dispatch as inputs to config loading so as to not require mutable access to the config after the repo has been loaded.

Reviewed By: quark-zju

Differential Revision: D53250254

fbshipit-source-id: a856adfbfe44aeb916cae435153677ab3bba978d