Pass field directly
Summary: Small refactor to pass in the FieldConfig instead of extracting values from it. This allows us to expand the field with new things and not need to plumb them through with additional props each time. It already combines 3 props into one, and one more would come in the next diff if not for this change.

Reviewed By: quark-zju

Differential Revision: D53330514

fbshipit-source-id: 5e2f8498d6bef0df50411341a5a3ec06c8cf9eab