clidispatch: allow marking command alias as "legacy" in Rust
Summary:
Add a way to mark a command alias as legacy in Rust.

I stuck it in the "aliases()" string to avoid adding another function that most commands don't care about. Maybe we should convert command definition metadata into more structured data.

As for "why", if we want to change the primary name of a command (e.g. "update" -> "goto"), we need to maintain backwards and forwards support.

For hooks, that means running command hooks using either the old or new name, so we need to know what the old name is.

Reviewed By: sggutier

Differential Revision: D53195188

fbshipit-source-id: fa8e0b3047fbf3e691fa1e281727dbfd15066bc7