clidispatch: invoke shell hooks for Rust commands
Summary:
Invoke "pre", "post", and "fail" hooks for native Rust commands.

We don't support Python hooks yet in Rust. If there are Python hooks configured, we fall back early the Python version of the command if possible, otherwise we will print a warning. There's no great way to know if you can fall back to Python for a given command, so I started with a hard-coded list of important commands we support in Rust which have a valid Python fallback.

As for args/environment setup, we support setting $HG_ARGS (list of raw CLI args), $HG (the binary location), $HG_RESULT for "post" hooks, and some telemetry related items. I didn't add support for $HG_PATS or $HG_OPTS since they didn't seem to be used anywhere, and are a bit more involved.

Reviewed By: quark-zju

Differential Revision: D53195195

fbshipit-source-id: 3b0fed15362801f5d910f616fb32f5b0e3ca57cb