hooks: fix "post" and "fail" hooks to shell escape args
Summary:
Similar to the "pre" command hook, the "post" and "fail" hooks now also shell escape the command args before joining on " ". This is so the hook can tell apart args with spaces.

I don't think any "post" or "fail" hooks use $HG_ARGS, but it's good to be consistent.

Reviewed By: quark-zju

Differential Revision: D53195194

fbshipit-source-id: acf9900b90a9b947e8bf358baf4afa310ee2390d