mercurial/bundles: migrate bundle creation to new bytes and futures
Summary:
This is diff 1/7 of a series that migrates the Mercurial wireproto from old bytes, futures, and streams to new ones.

In this diff, we focus on the bundle creation process.   Previously this used a state machine to handle the sequential nature of bundle generation, but we can replace this with clearer code using async-await syntax.

Reviewed By: quark-zju

Differential Revision: D51303208

fbshipit-source-id: efce817b61784d084849df8a5da4482d717197e2