walker: remove use of async-compression
Summary:
This currently uses a metered stream to measure the amount of compressed data.  However, this is overkill, as we only write everything at once, and always buffer the full compressed data, so we can simplify and just ask `zstd` to compress it for us and take the length of the compressed data.

This removes a dependency on `async-compression`, which uses old-style streams.

Reviewed By: quark-zju

Differential Revision: D51303200

fbshipit-source-id: 2014e2576f92bdc23233ed0255cc2c6b1a8e1d7d