hgproto: migrate bundle decoding to new bytes and streams
Summary:
This is diff 7/7 of a series that migrates the Mercurial wireproto from old bytes, futures, and streams to new ones.

This is the last step in removing all of old bytes and streams from wireproto handling.  Using the components we migrated in the rest of the stack, we fully update the bundle decoder and Mercurial protocol handler to use the new types.  This needs to be done in one go as it's not possible to change incrementally: the types need to agree throughout.

Reviewed By: quark-zju

Differential Revision: D53266420

fbshipit-source-id: 364b28a8091fff06d63eaee5816c408a0f2f9bc0