megarepo: support phabricator diff ids
Summary:
Make megarepo and fbcodereview extensions aware of each other so megarepo can work for cross-repo resolution of phabricator diff IDs.

In particular, fbcodereview extension now doesn't abort the the command when the phabricator diff is associated with a different repo. This gives the megarepo extension a chance to convert it in a later "autopull" attempt.

Reviewed By: quark-zju

Differential Revision: D53432226

fbshipit-source-id: 1eb5ae0746404c521a580c8bf94a5a7cc6a9e43b