Fill commit message actions
Summary:
Add actions to fill an empty commit message from another source. Here I add buttons for loading from the previous commit or the commit message template.

Since this overwrites your commit message, we should take care to only do this when it can be applied without overwriting anything.

So we check if the newly loaded message has any fields which conflict. If none conflict (either empty or identical), then the new message is merged.

If any fields conflict, we show an interstitial asking to confirm, and listing out the conflicts.

Reviewed By: quark-zju

Differential Revision: D53365522

fbshipit-source-id: 19774ad56e8019a835e48bb7606c8fc0b75f82fd