Add support for submenus in context menus
Summary:
Add a new `type` for menu items in context menus, which will show a submenu. This is common in context menus. The submenu is styled just like the original menu.

Details:
- Each "level" of menu may only be showing one submenu at a time
- hovering on a context menu item which is of type `submenu` will start showing the submenu on the right side
- mouse exiting the submenu or entire menu does not hide any submenus. You can move your mouse around outside the context menu freely
- hovering on any menu item clears any submenus at that level. So, if you have a submenu open, then hover a different item at the original level, the submenu will disappear.
- Any additional submenus hide when the parent submenu hides.

This UX is actually simple to achieve by having onMouseEnter of a submenu store and remember that submenu as active. onMouseEnter any other item clears the submenu. Then, each level of context menu renders its submenu. So clearing a submenu implicitly clears/hides all children subments. We compute the position of the submenu at mouse enter time based on the target and parent.

Reviewed By: quark-zju

Differential Revision: D53364448

fbshipit-source-id: 46036beaed52dadf14012a8b719a00c375ccb675