Use TypedEventEmitter instead of importing from events
Summary:
We saw an issue with `sl web` when running as a webview on macos. Though we didn't have source maps, it seems the error was instantiating `EventEmitter` from `events`. This must be polyfilled to use in the browser (it's a nodejs thing). It seems some layer of that polyfill is not working in some build environments...

We should instead use our TypedEventEmitter which does not depend on importing `events` (it's entirely our code).

Reviewed By: quark-zju

Differential Revision: D53452421

fbshipit-source-id: 53ebcda2f75deda088e9f1ddea931d6102be3646