configloader: record files that don't exist as "loaded"
Summary:
When loading a config file and the file doesn't exist, we now track that file in `files`.

Python uses the `files` to determine if the config in-hand has loaded the repo's config or not. In the case where the repo had no config file we were erroneously concluding the repo config had not been loaded.

Reviewed By: quark-zju

Differential Revision: D53369768

fbshipit-source-id: 91a9de5e89a5637826073447be24f492f87e01f4