Close other 'click' tooltips when opening a new one
Summary:
If you click the settings button to open the dropdown, then the bug button, it didn't hide the settings dropdown. This has been a minor pet peeve of mine for a while now :P

With this change, tooltips can specify a "group" they belong to. Opening any tooltip in a group closes all other tooltips in the group.

We use this to have every dropdown in the top bar close all other dropdowns.

Is the "group" abstraction necessary? We could also choose to just have this happen automatically for any "click" tooltip.

Reviewed By: quark-zju

Differential Revision: D53458664

fbshipit-source-id: 4df63ae008575a910ef51f0edaa3abbe6886a0b1