commit: fix commit (and amend) -i not picking up changes over filemode
Summary:
`sl commit` and `sl amend` usually pick up changes to file modes when not run interactively. This fixes a bug where this kind of changes were not picked up if running `commit --interactive` or `amend --interactive`

`special_re` is used to see whether some patch/change is to be considered "special" at the time of patching. This is also how in `patch.py` it's determined whether a new file was added or an `hg mv` was performed:

```
        sum(
            [h for h in pycompat.itervalues(applied) if h[0].special() or len(h) > 1],
            [],
        ),
```

Reviewed By: muirdm

Differential Revision: D53455098

fbshipit-source-id: 1963be452ad6068a4665597ddf57a495ac94a74f