mononoke_queries: don't cache if query returns empty rows in auto caching
Summary:
Auto caching currently caches whatever `cacheable` marked queries return even if they return empty rows. Negative caching is tricky to implement right and we generally try to avoid it in Mononoke, and one particually annoying consequence of this is that it breaks the commonly used pattern of first querying something from a replica and if not found querying master, because the first query to the replica gets cached.

This affects `SelectBubbleById` query in the `EphemeralStore` that currently fails when querying for a bubble id that hasn't yet reached the queried replica.

Reviewed By: markbt, liubov-dmitrieva

Differential Revision: D53353481

fbshipit-source-id: 7496d6f6781f7606a00f1d6ffbc49629b4b61830