Add integration tests for expansion of git submodules
Summary:
Adding an integration test to cover the basic use case along with recursive submodules.

The test sets up 3 git repos: A, B and C. A depends on B (as a submodule), which in turn depends on C.

I import each repo into Mononoke then run the forward syncer initial import command. Then, I make changes on the submodule repos and do the same.

The new functionality to expand submodules isn't implemented yet, so this is just doing the default behaviour of stripping the submodules.
But I figured it would make reviewing much easier if I set up the integration test in a diff before.

After the expand submodules functionality is implemented, the last `tree` output should become:

```
  $ tree
  .
  `-- smallrepofolder1
      |-- git-repo-b
      |   |-- bar
      |   |   `-- zoo
      |   |-- foo
      |   `-- git-repo-c
      |       |-- choo
      |       |-- choo3
      |       |-- choo4
      |       `-- hoo
      |           `-- qux
      |-- regular_dir
      |   `-- aardvar
      `-- root_file
```

Which basically means that submodule contents are copied into the small repo directory recursively.

There is one improvement that I want to make to this test later: set up an automatic live sync with the forward syncer, instead of running the `initial-import` command again. I tried doing that, but didn't get it working, so decided to move on and come back to it later.

Reviewed By: mitrandir77

Differential Revision: D53310533

fbshipit-source-id: c5065e50c998019cc7e746b79a6d0f3a13d0b641