stop special treatment of the root nodes
Summary:
Across the walker codebase we have a few places where we trigger
checks for whether node should be visited like:
 * is the node already visited?
 * is the type of edge excluded from the walk
 * is the type of edge on the always_emit list

In addition to those in few places we force allowlist traversing of walk root. I think we can get away with not doing that and the walker will keep working as usual (or at least the integration tests pass!).

My ulterior motive here is to centralize the logic of what's visited in "needs_visit" function. This way I'll be able to exclude some nodes from being visited just by modifying that function.

But.. some of those unnecessary "if's" are, I think, performance optimizations. So let's be on watch for the next few days for any signs of walker being slower.

Differential Revision: D53501621

fbshipit-source-id: 9ef8732da018ae7fd3dcb987050d8faf097b1ce2