add --exclude-node feature
Summary:
We have some corrupted commits in Mononoke that should never be traversed by the walker. It would simply crash on them.

I thought about whether this particular problem should be ignored by walker but it doesn't sound that good to me: I don't want the walker to ignore problems with parsing bonsais or hg changesets in the future. The only other way to deal with the issue seems to be denylisting those changesets during walk.

The flag I'm adding here makes it easy to skip any "Node" during walk - this way we can prepare ourselves for future cases of data coruption.

Differential Revision: D53501620

fbshipit-source-id: e50148d5e9f45c4c493821236bf4b33c11feb738