use FetchMode instead of local_only in backingstore layer
Summary:
This diff changes the Rust bindings to accept an enum instead of a boolean flag. This makes a follow up bootcamp task a bit easier in adding a `RemoteOnly` option.

I didn't do much from the C++ side, as I plan to make that refactor a part of the bootcamp task. This was just to unblock the ffi stuff

Reviewed By: muirdm

Differential Revision: D53362847

fbshipit-source-id: d809c5406a067eb1dbfe97c548787e511b95dfbc