derived data: Derive dependent types in batch derivation
Summary:
It makes sense that backfill_exactly_batch requires for all ancestors of the batch
to have been derived:
That's important to ensure we don't have an unbounded traversal to do to make sure
we reach that state.

However, dependent data types are very limited (the sdepth of the dependency graph is 3 at the moment),
and requiring that dependent derived data types are already derived basically forces the caller to
manually perform the graph resolution.

Remove the undocumented requirement that dependent derived data types must be derived first.

Reviewed By: singhsrb

Differential Revision: D53530526

fbshipit-source-id: 5ceedb7c04ebd78364b26b5a7ce8c0b687734cca