derived data: Implement batch derivation for dependent derived data types
Summary:
This is a step towards enabling batch derivation of multiple data types, which will make the batch derivation
API much more usable.

In particular, this will allow us to use it to backfill derived data in gitimport, but will also help us
write a modern implementation of `backfill_derived_data` without having to do a number of hacks and
sacrificing the UX as is currently the case.

**A note on the implementation:**

Derived data dependencies are declared at compile time by populating the associated type: `Dependencies`
on `BonsaiDerivable`.

This associated type must implement the `DerivationDependencies` trait.

This is nice, because it is correct by construction. Types are the source of truth for these dependencies.

On the other hand, walking the graph must be done by implementing the `DerivationDependencies` trait, which
is slightly cumbersome.
I can only guess this is why this type-safe dependency graph is vastly under-used and currently only
used to check whether dependencies are derived.

Other places in the code rely on `DERIVED_DATA_ORDER` which is a manually hard-coded version of the dependency
graph between the derived data types as a `static HashMap`, but we want to eventually remove it as it is a point of
duplication which needs maintaining every time derived data types get created or deleted.

Reviewed By: singhsrb

Differential Revision: D53530521

fbshipit-source-id: fe2bf4c2238a3fbca2b16b91d1376eae83498ad6