derived data: Expose backfill batch behaviour in integration test
Summary:
If a derived data type depends on another data type, (for instance, blame depends on unodes),
Trying to derive data in a batch will fail if the dependent type wasn't already derived.

Document this in a new integration test as I want to change this behaviour later in this stack.

Reviewed By: singhsrb

Differential Revision: D53530524

fbshipit-source-id: 4c9e9dc80ed3dde743ef5451f9179f40de78d070