Eden cli: On rm, check if the arg is a file instead of a directory
Summary:
During my first test of rm I saw that if rm is passed a file (that exists) instead of a directory the message is unclear (correct, but not obvious if you don't realize you passed the wrong thing).

I think just reporting that it's a file and not a directory and still allowing to delete it is a bit better.

Reviewed By: MichaelCuevas

Differential Revision: D53551256

fbshipit-source-id: 1ec367267d53185dec5ddac69d44039e39ea743c