Collect rage in bug report
Summary:
Collect an hg rage when reporting a bug.

In the next diff I'll add a checkbox to the UI so this is optional (since it takes an extra 60s+ to collect!).

This will be nice to have to debug certain types of issues more easily without having to ask users to collect a rage after they repro next time.

Also, for some users, this might be a convenient way to ask them to collect a rage with a GUI instead of a CLI. E.g. windows ISL webview.

Reviewed By: quark-zju

Differential Revision: D53588451

fbshipit-source-id: d6b3376f10e0810aaaeb9134cca220f16b93ed0b