Migrate OperationDisabledButton to jotai
Summary:
Migrate this button. This one was an atomFamily, which works just slightly differently.

Technically, this has a memory leak, since some OperationDisabledButtons have unique keys per hash they apply to. But this shouldn't be a big deal. We also had this leak in recoil. At least now we're cognizant of it!

Reviewed By: quark-zju

Differential Revision: D53537127

fbshipit-source-id: 86519601fe859860d66bf919b4b596e8ddc55258