Migrate eden du to getScmStatusV2
Summary:
# Context
The "eden du" command displays the actual storage used on disk. This is used when users want to know how much actual storage is being used by eden for purposes such as clearing up disk space.

# This Diff
This diff migrates the thrift endpoint from getScmStatus to getScmStatusV2

# Technical Details
getScmStatusV2 packages the previous values of getScmStatus into a substruct and adds some new field. This diff moves the previous values of getScmStatus into that substruct and uses default values for the fields not previously present.

Reviewed By: MichaelCuevas

Differential Revision: D53581822

fbshipit-source-id: 7fcb756d5096a2a0a53075870bae979a301f0e87