Optionally limit the number of PRs created in each submission (#824)
Summary:
Optionally limit the number of PRs created in each submission

Currently, when submitting a PR in Sapling, it's easy to accidentally create multiple PRs when only one was intended (e.g. forgetting to run `sl pr follow`, or editing a PR that is based off a different branch).

With this change, you can configure Sapling to limit the number of PRs that can be created in a single submit action, e.g. to just 1:
```
sl config --user github.max-prs-to-create=1
```
Now when attempting to create multiple PRs in a single `sl pr submit` action, Sapling will fail with:
```
abort: refused to create 2 pull requests, max is 1
```
(Note that this setting doesn't limit the size of a PR stack: a larger PR stack can be made by submitting each new PR separately. Re-submitting the top PR will still update the entire stack below it, no matter the size.)

Pull Request resolved: https://github.com/facebook/sapling/pull/824

Reviewed By: quark-zju

Differential Revision: D53010053

fbshipit-source-id: d39e0665ffadf6794ce6cc4458f22d24b8662b1c