Make getUiState message not platform-specific
Summary:
Our vscode webview is currently broken, due to the platform not getting properly set up ahead of other code.

I tracked this down to the vscode platform definition including AccessGlobalRecoil, which recently started importing from jotai, which includes a ton of imports, causing it to include ClientToServerAPI and eventually urlParams, which checks the platform statically.

To fix this, we can untangle this knot simply by extracting the functionality that used AccessGlobalRecoil. There's no real reason for it to be vscode-only. Any platform could want UI state, and I see no problem with that.

This way, our vscode entry point is a much tighter set of dependencies. The only "scary" one remaining is vscode/Internal, but right now that doesn't have any other notably imports.

Reviewed By: muirdm

Differential Revision: D53619177

fbshipit-source-id: d3c6de16af4001b62f5977905c06176b76b1d069