ComparisonView: resolve React act warning in tests
Summary:
Some tests (large diffs) end with the syntaxHighlighting still running.
While `cleanup` is used that might eventually call `token.cancel()` to
cancel the running tasks. The `cleanup` seems to be async and does not
prevent the warning:

    Warning: An update to null inside a test was not wrapped in act(...).
    ...
      53 |     tokenizeHunks(theme, path, hunks, token).then(result => {
      54 |       if (!token.isCancelled) {
    > 55 |         setTokenized(result);
         |         ^
      56 |       }
      57 |     });
      58 |     return () => untrackCancellationToken(token);
      at src/ComparisonView/SplitDiffView/syntaxHighlighting.tsx:55:9

Fix it by adding a sync `cancelAll` function that tests can explicitly
use in their `sync` logic. Note this cannot be `afterEach` since that
also seems to be async-ish.

Reviewed By: evangrayk

Differential Revision: D53622815

fbshipit-source-id: 0d497d464aac127ca167f5cd83030cfdeaca792b