workingcopy: make watchman client shareable
Summary:
Fiddle with things so watchman client lives on the WorkingCopy. I'm going to use it to implement the "hgevents" functionality which happens even in EdenFS checkouts (so the watchman client needs to live someplace more common).

We still initialize the watchman client lazily, so we won't try connecting if watchman isn't needed.

Reviewed By: quark-zju

Differential Revision: D53553929

fbshipit-source-id: 9e6808e6b1d0e2b3c8fd35390afc36eaabaf84f5