dispatch: propagate context from Rust to Python
Summary:
Now we pass a context object instead of just the config (the context contains the config).

The uiconfig stores the context and config separately, now. The config is separate so it is mutable, and we remember the context so we can pass it back into Rust as needed.

When passing back into Rust, we want to incorporate any config mutations made in Python, so I added a `ui.rustcontext()` method that swaps the original context's config with the current uiconfig._rcfg.

Reviewed By: quark-zju

Differential Revision: D53553928

fbshipit-source-id: 132ea25126f2b309361c8564131746768a2a54a3