manual mirror of commitByHash
Summary:
This selector family is used in a lot of places. But it's challenging to migrate all users at once, so let's just duplicate it with a second cache. In this case, it's not very harmful to just have a second cache of this.

This is very temporary anyway, the original should go away in this stack to re-consolidate.

Reviewed By: quark-zju

Differential Revision: D53632285

fbshipit-source-id: 044a32fd929605e5e6a3cf59e180e8c84e0948bd