derived data: Remove gap_size option from the batched derivation interface
Summary:
This feature was used to speed-up backfilling of some derived data types at the cost of
leaving gaps (only deriving a subset of the changesets).

Data derivation performance has improved a lot since it was deemed necessary.

Using this reduces predictability of the state the system is in.

Let's be opinionated and say goodbye: It removes a good amount of code and makes the system less complex/more predictable.

Reviewed By: markbt

Differential Revision: D53618305

fbshipit-source-id: 9eea0833313eb3cb8631280a1bea820158fd06c0