edenapi: disable min speed transfer check for land_stack()
Summary:
Similar to D42478560, this diff disables min speed check for land stack requests. For land stack requests, server returns the result once it is fully ready, and it might be slow if the stack is large, so the check needs to be disabled.

We still have a general timeout for an edenapi request.

Reviewed By: markbt

Differential Revision: D53665467

fbshipit-source-id: 733757aef550694282c5af865725fafdb895eb15