Use local cache for avatars if localStorage not working
Summary:
It seems our migration to jotai uncovered a werid case in our avatar fetching. If localstorage is not working for some reason, then it will keep trying to fetch avatars in a loop endlessly, destroying performance. I think that in vscode, sometimes localstorage isn't reliable. We may need to switch this to use a vscode API instead of just the browser one. Other platforms like androidStudio don't have localstorage implemented either, so we should not be assuming this can always cache stuff.

We should probably fix our localstorage API abstraction so this is not an issue. In the mean time, we should cache avatars locally as well, so that we can rely on the avatars being cached and not refreshed.

Reviewed By: quark-zju

Differential Revision: D53671114

fbshipit-source-id: 219175e07abce77c145acdc2c29d0d6cb1b2a3f4