jotaiUtils: add atomFamilyWeak to avoid memory leak
Summary:
The initial motivation of using Jotai is to avoid memory leak caused by
`selectorFamily` etc. But the `atomFamily` in `jotai/utils` also has the memory
leak issue. This diff adds a drop-in replacement that tries to mitagate the
memory leak problem. The idea is that by using weakRef and periodical cleanup,
the wasted memory is bounded to be about 2x the needed memory, and the runtime
overhead is also bounded at 2x worst case (or 4x, if clearing strong
cache does not immediately release objects in the weak cache).

It has a "regular" `cache` and a "weakRef" cache. The "weakRef" cache ensures
the `atomFamilyStateless` (see below) use-cases are covered - if an atom is in use then
it is kept alive. The "regular" cache is for the memory-leak `atomFamily` use-case
where the cache can be used across component unmount/remounts.

There is another interesting pattern for some of the `useStrongCache = false`
cases (initially I wrote this as `atomFamilyStateless`):

    // define atomFamily as pure function
    const family = (k) => atom(...);

    // in React component, use it together with useMemo
    const memoAtom = useMemo(() => family(k), [family, k]);
    const value = useAtomValue(memoAtom);

It solves the memory leak problem more "perfectly" with some downsides:
- Without `useMemo` it will infinite loop re-render with React StrictMode.
- Atom caching is per component, not shared across different components.
- When beinging `get`-ed in other atoms, it might create tempoary atoms
  that looks less elegant.
That said, this pattern does seem more elegant if:
- No dependent atoms that use the family atoms (not used in `get`s).
- Only 1 component uses `useAtomValue` for this state.

Reviewed By: evangrayk

Differential Revision: D53643859

fbshipit-source-id: 3995ab90aaa748232dc8e093c7f2dcd0eed085d8