hideOperation.test: migrate to DAG renderer
Summary:
There is a behavior difference that the "public base" is not hidden during
optimistic view. To implement this behavior we might want to change the
"subset" to render to also consider skipping public commits like how
"obsoleted" commits get skipped. The hide operation itself should not modify
the DAG to remove public commits.

Reviewed By: evangrayk

Differential Revision: D53668273

fbshipit-source-id: df12dba2958c53f73de2e585e0d6141e9bafa6c9