isl: handle isl upgrade more smoothly on Windows
Summary:
The `isl` command wants to extract ISL source to a directory. It will rename
the outdated directory and delete it. The rename step can error out if the
rename destination exists and cannot be deleted (on Windows this happens).
Handle this case by moving to a sub-directory of the "to delete" directory
instead.

Reviewed By: muirdm, evangrayk

Differential Revision: D53710946

fbshipit-source-id: e64106a696ddf745263ab8edef1ea274a44e16fd