eden cli: Refactor handler.exe code a bit
Summary:
This diff moves the code that manages handler.exe to its own module and abstract the functionality into a base class, even though for now only Windows is implemented.

It also calls this new handle in the branch that deletes a non-Eden directory, because it's very convenient to test.

Reviewed By: genevievehelsel

Differential Revision: D53594227

fbshipit-source-id: 309efab3f89893f745a82fe5acd128ecd702d42e