git: Fix failure to round-trip when generating git commit
Summary:
Before this diff, we would parse "John Doe <john.doe@gmail.com>" as
* name: "John Doe "
* email: "john.doe@gmail.com"

Note the trailing space in the name.

This means that deriving a git imported commit would create a git with a different hash from the input.

We didn't hit it because for imported commits, we rely on the mapping.

Reviewed By: mzr

Differential Revision: D53714717

fbshipit-source-id: 4b9b34c7a9829c6fa4906ffa9651be93ca03cb56