make test non-flaky
Summary:
This test relies on timing which is common source of flakiness. I don't
eliminate timing but I make the treshold more reasonable as in this test we're
trying to distinguish between something that should be pretty much instant and
something that would take ~900ms.

In `dev-lg` mode the instant bit takes a little bit longer than 100ms. I've set
the treshold to 600ms which it should never hit.

Reviewed By: markbt

Differential Revision: D53729116

fbshipit-source-id: 8a509eebdf120313d8299dfb59d75666fe3b9690