derived data: Declare the dependency between bssm and skeleton manifest
Summary:
The `dependencies` macro is currently one of the many places where the dependency graph of
derived data types is defined.

Thanks to this duplication, we got away without this declaration in the past.

As I refactor to consolidate the implementations around this type-safe one, it will become
crucial that this is correct.

Reviewed By: YousefSalama

Differential Revision: D53668709

fbshipit-source-id: 5646d313c9d418d6b58bbaf3bd0ec19dfa9d7950