derived data: Remove perform_single_derivation
Summary:
This function is conceptually fully duplicated with `derive_exactly_batch`, except it doesn't provide the functionality
of deriving more than one csid at a time.

In practice, it also did a couple more things:
* Check whether the changeset `needs_rederive` and rederive it in that case
* Log the discovery stats to scuba

These behaviours were moved to the only caller.

A couple of TODOs were added for a couple of improvements we could make. I'm not adding them to this diff to avoid mixing
refactoring and behavioural changes.

Reviewed By: YousefSalama

Differential Revision: D53666356

fbshipit-source-id: d9955a07a94a1f775113d094232b7383f895660d