implement --invoke-command in Plugboard CLI
Summary:
This is an early diff that sketches out the most bare minimum happy
path. Flesh out the invoke_command Thrift endpoint and add an
--invoke-command CLI option. It looks up a command, fetches its
source, and invokes the "starlark" pipeline.

In the future, we should decouple command execution from pipelines and
have this Thrift endpoint and the Starlark operator call into the same
library.

Also, to increase reliability and tail latencies, we should cache
command source code in-process.

Differential Revision: D53733586

fbshipit-source-id: 1f82308706f1a023a38d51ec8a7d69d9b3686097