serverAPIState: remove recoil usage
Summary:
This is the last state in ISL Recoil.

I had to change `simulateServerStatusChange` to also call `simulateServerStatusChange`.
I believe the old code happened to work because the `reconnectingStatus` is
updated when registering the callback (`messageBus.onChangeStatus`), and the test
does not actually trigger the callback later. Now with Jotai we only register
the callback once (unless hot reload) we need the change callback to be
triggered after registration.

Reviewed By: evangrayk

Differential Revision: D53781476

fbshipit-source-id: c7a98d51444c03f9a36a93d65c4cf1dae607b57e